Check botocore changelog for service updates
