+++ docs/MODULES.md
@@ -85,7 +85,7 @@ Full list of [mypy-boto3-builder](README.md#mypy_boto3_builder) project modules.
         - [Syntax Type Map](mypy_boto3_builder/type_maps/syntax_type_map.md#syntax-type-map)
         - [Typed Dicts](mypy_boto3_builder/type_maps/typed_dicts.md#typed-dicts)
     - [Utils](mypy_boto3_builder/utils/index.md#utils)
-        - [Boto3Changelog](mypy_boto3_builder/utils/boto3_changelog.md#boto3changelog)
+        - [BotocoreChangelog](mypy_boto3_builder/utils/botocore_changelog.md#botocorechangelog)
         - [Markdown](mypy_boto3_builder/utils/markdown.md#markdown)
         - [NicePath](mypy_boto3_builder/utils/nice_path.md#nicepath)
         - [PyPIManager](mypy_boto3_builder/utils/pypi_manager.md#pypimanager)
